+++ HISTORY.rst
@@ -12,8 +12,13 @@ Future History
 Recent History
 ~~~~~~~~~~~~~~
 
-.. Next Release
-.. ------------------
+Next Release
+------------------
+
+* Turn on waitlist with manual processing.
+* Collect guest counts, but no personal information.
+* Throttle Wild Apricot requests.
+* Allow members to see the list of event registrants.
 
 0.2.0 (2020-11-08)
 ------------------